backend/requirements.txt
google-generativeai
flask
pinecone
groq
python-dotenv
